refactor : Made wager page a component
